fixes #52532
A rule like this would allow 10 jpg files paralell

 {
  "domainPattern" : null,
  "filenamePattern" : ".+\\.jpg$",
  "accountPattern" : null,
  "enabled" : true,
  "maxSimultanDownloads" : 10,
  "allowToExceedTheGlobalLimit" : true
}

Please note that I renamed the pattern field to domainPattern. the old getter will convert old fields to the new one. let's keep it for a couple of months.


Improved the Flickr Decrypter. It tries to set a proper filename now.